chore(settlement): rename api to settlement_api to satisfy clippy
