+++ crates/ilp-node/src/node.rs
@@ -30,7 +30,7 @@ use interledger::{
         EchoService, ExchangeRateFetcher, ExchangeRateService, ExpiryShortenerService,
         MaxPacketAmountService, RateLimitService, ValidatorService,
     },
-    settlement::api::{create_settlements_filter, SettlementMessageService},
+    settlement::settlement_api::{create_settlements_filter, SettlementMessageService},
     store_redis::{Account, AccountId, ConnectionInfo, IntoConnectionInfo, RedisStoreBuilder},
     stream::StreamReceiverService,
 };
