+++ crates/interledger-settlement/src/settlement_api/message_service.rs
@@ -121,8 +121,8 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::api::fixtures::{BODY, DATA, SERVICE_ADDRESS, TEST_ACCOUNT_0};
-    use crate::api::test_helpers::{block_on, mock_message, test_service};
+    use crate::settlement_api::fixtures::{BODY, DATA, SERVICE_ADDRESS, TEST_ACCOUNT_0};
+    use crate::settlement_api::test_helpers::{block_on, mock_message, test_service};
     use interledger_packet::{Address, Fulfill, PrepareBuilder, Reject};
     use std::str::FromStr;
     use std::time::SystemTime;
